+++ pkg/reconciler/kubernetes/adapter.go
@@ -329,3 +329,8 @@ func toResource(m *internal.Metadata) *Resource {
 		Namespace: m.Namespace,
 	}
 }
+
+func ToUnstructured(manifest []byte, async bool) ([]*unstructured.Unstructured, error) {
+	// expose the internal unstructured converter
+	return internal.ToUnstructured(manifest, async)
+}
