+++ pkg/reconciler/kubernetes/internal/test/manifest.yaml
